Fix Currency Amount Parsing

+++ ProsperousUniverse.API/DTOs/CurrencyAmountDTO.cs
@@ -10,7 +10,7 @@ public sealed class CurrencyAmountDTO
     [GraphQLNonNullType]
     public string Currency { get; set; }
 
-    public int Amount { get; set; }
+    public double Amount { get; set; }
 
     public static CurrencyAmountDTO Parse(JsonElement jsonElement)
     {
@@ -18,7 +18,7 @@ public sealed class CurrencyAmountDTO
         {
             Currency = jsonElement.GetProperty("currency").GetString() ??
                        ParsingUtils.ThrowRequiredPropertyMissing("currency"),
-            Amount = jsonElement.GetProperty("amount").GetInt32()
+            Amount = jsonElement.GetProperty("amount").GetDouble()
         };
     }
 }
